extract rubric form and related functions to own component
-#330
