+++ frontend/src/hooks/useTemplate.ts
@@ -1,5 +1,6 @@
 import { useState } from "react";
 import { Template } from "palette-types";
+import { useFetch } from "./useFetch.ts";
 
 export const useTemplate = () => {
   // this template tracks the template that is currently being updated
@@ -13,6 +14,11 @@ export const useTemplate = () => {
   );
   const [templateInputActive, setTemplateInputActive] = useState(false);
 
+  const { fetchData: putTemplate } = useFetch("/templates", {
+    method: "PUT",
+    body: JSON.stringify(importingTemplate),
+  });
+
   return {
     updatingTemplate,
     importingTemplate,
@@ -20,5 +26,6 @@ export const useTemplate = () => {
     setImportingTemplate,
     setUpdatingTemplate,
     setTemplateInputActive,
+    putTemplate,
   };
 };
